+++ mtg_proxy_printer/model/carddb.py
@@ -338,7 +338,7 @@ class CardDatabase(QObject):
         ''')
         if order_by_print_count:
             query += '    LEFT OUTER JOIN LastImageUseTimestamps USING (scryfall_id, is_front)\n'
-        where_clause = ['WHERE Printing.is_hidden IS FALSE']
+        where_clause = ['    WHERE Printing.is_hidden IS FALSE']
         where_parameters = []
         if card.language:
             where_clause.append(f'AND "language" = ?')
@@ -410,6 +410,51 @@ class CardDatabase(QObject):
                 is_oversized, face_number, language in cursor
         ]
         return result
+
+    def find_related_cards(self, card: Card) -> CardList:
+        """
+        Recursively finds all cards related to the given card.
+        This may be cards referenced by name in either direction, or token cards created
+        """
+        query = cached_dedent("""\
+        WITH RECURSIVE 
+          source_oracle_id (card_id) AS (
+            SELECT card_id
+            FROM Card
+            WHERE oracle_id = ?),
+          related_oracle_ids(related_id) AS (
+            SELECT related_id
+              FROM RelatedPrintings
+              JOIN source_oracle_id USING (card_id)
+            UNION
+            SELECT RelatedPrintings.related_id
+              FROM RelatedPrintings
+              JOIN related_oracle_ids ON RelatedPrintings.card_id = related_oracle_ids.related_id
+              WHERE RelatedPrintings.related_id NOT IN (SELECT source_oracle_id.card_id FROM source_oracle_id)
+        )
+        SELECT oracle_id
+          FROM Card
+          JOIN related_oracle_ids ON Card.card_id = related_oracle_ids.related_id
+        """)
+        related_card_ids = self.db.execute(query, (card.oracle_id,)).fetchall()
+        cards = []
+        for related_oracle_id, in related_card_ids:
+            # Prefer same set over other sets, which is important for multi-component cards like Meld cards. If it
+            # isn't available, take from any other set. As a last-ditch fallback, resort to English printings.
+            # The last case is most likely hit with non-English token-producing cards,
+            # as long as Scryfall does not provide localized tokens.
+            related_cards = self.get_cards_from_data(
+                CardIdentificationData(card.language, set_code=card.set.code, oracle_id=related_oracle_id),
+                order_by_print_count=True) or \
+            self.get_cards_from_data(
+                CardIdentificationData(card.language, oracle_id=related_oracle_id),
+                order_by_print_count=True) or \
+            self.get_cards_from_data(
+                CardIdentificationData("en", oracle_id=related_oracle_id),
+                order_by_print_count=True)
+            if related_cards:
+                cards.append(related_cards[0])
+        return cards
 
     def find_collector_numbers_matching(self, card_name: str, set_abbr: str, language: str) -> StringList:
         """
@@ -693,8 +738,9 @@ class CardDatabase(QObject):
             bool(highres_image), bool(is_oversized), face_number
         )
 
-    def store_current_printing_filters(self, use_transaction: bool = True, *, force_update_hidden_column: bool = False,
-                                       progress_signal: typing.Callable[[], None] = None):
+    def store_current_printing_filters(
+            self, use_transaction: bool = True, *,
+            force_update_hidden_column: bool = False, progress_signal: typing.Callable[[], None] = None):
         if progress_signal is None:
             progress_signal = (lambda: None)
         section = mtg_proxy_printer.settings.settings["card-filter"]
@@ -787,7 +833,6 @@ class CardDatabase(QObject):
             WHERE Printing.is_hidden IS FALSE
           );
         """))
-
         progress_signal()
         # Performance note: Using INSERT OR IGNORE and removing the inner scryfall_id NOT IN (subquery) simplifies the
         # query plan, but takes about 40% longer to evaluate (on the card data of late April 2022)