+++ mtg_proxy_printer/sqlite_helpers.py
@@ -87,7 +87,7 @@ def open_database(
     db = sqlite3.connect(db_path, check_same_thread=check_same_thread)
     logger.debug(f"Connected SQLite database {location}.")
     # These settings are volatile, thus have to be set for each opened connection
-    db.executescript("PRAGMA foreign_keys = ON; PRAGMA analysis_limit=1000; PRAGMA trusted_schema = OFF;")
+    db.executescript("PRAGMA foreign_keys = ON; PRAGMA trusted_schema = OFF;")
     logger.debug("Enabled SQLite3 foreign keys support.")
     if should_create_schema:
         populate_database_schema(db, schema_name)