+++ tests/test_carddb.py
@@ -749,10 +749,41 @@ def test_is_removed_printing(
 ])
 def test_get_basic_land_oracle_ids(
         qtbot, card_db: CardDatabase,
-        include_wastes: bool, include_snow_basics: bool, expected_oracle_ids: typing.List[str]):
+        include_wastes: bool, include_snow_basics: bool, expected_oracle_ids: StringList):
     fill_card_database_with_json_cards(
         qtbot, card_db, ["english_basic_Forest", "english_basic_Wastes", "english_basic_Snow_Forest"])
     assert_that(
         card_db.get_basic_land_oracle_ids(include_wastes, include_snow_basics),
         contains_inanyorder(*expected_oracle_ids)
+    )
+
+
+@pytest.mark.parametrize("source_id, expected_cards_names", [
+    ("2c6e5b25-b721-45ee-894a-697de1310b8c", ["Food"]),  # Bake into a Pie
+    ("37e32ba6-108a-421f-9dad-3d03f7ebe239", []),  # Food token
+    ("e4b7e3b5-2f3c-4eb7-abc9-322a049a9e1a", []),  # Food Token
+    # Both printings of Asmoranomardicadaistinaculdacar
+    ("d99a9a7d-d9ca-4c11-80ab-e39d5943a315", ["The Underworld Cookbook", "Food"]),
+    ("2879f780-e17f-4e68-931e-6e45f9df28e1", ["The Underworld Cookbook", "Food"]),
+    # The Underworld Cookbook
+    ("4f24504e-b397-4b98-b8e8-8166457f7a2e", ["Asmoranomardicadaistinaculdacar", "Food"]),
+
+])
+def test_find_related_printings(qtbot, card_db: CardDatabase, source_id: str, expected_cards_names: StringList):
+    fill_card_database_with_json_cards(
+        qtbot, card_db, [
+        "The_Underworld_Cookbook",
+        "Food_Token",
+        "Asmoranomardicadaistinaculdacar",
+        "Bake_into_a_Pie",
+        "Asmoranomardicadaistinaculdacar_2",
+        "Food_Token_2",
+    ])
+    source_card = card_db.get_card_with_scryfall_id(source_id, True)
+    assert_that(source_card, is_(not_none()))
+    related = card_db.find_related_cards(source_card)
+    assert_that(
+        related, contains_inanyorder(
+            *[has_property("name", equal_to(expected)) for expected in expected_cards_names]
+        )
     )